feat(Grid): Migrate to ts
